create PlayerControlButton component
