+++ src/shared/components/VideoPlayer/VideoPlayer.style.tsx
@@ -3,6 +3,9 @@ import styled from '@emotion/styled'
 
 import { SvgPlayerSoundOff } from '@/shared/icons'
 
+import { PlayerControlButton } from './PlayerControlButton'
+import { ControlButton } from './PlayerControlButton.style'
+
 import { colors, sizes, transitions, zIndex } from '../../theme'
 import { Text } from '../Text'
 
@@ -13,11 +16,6 @@ type ContainerProps = {
 type CustomControlsProps = {
   isFullScreen?: boolean
   isEnded?: boolean
-}
-
-type ControlButtonProps = {
-  showTooltipOnlyOnFocus?: boolean
-  disableFocus?: boolean
 }
 
 const focusStyles = css`
@@ -61,97 +59,6 @@ export const CustomControls = styled.div<CustomControlsProps>`
   z-index: ${zIndex.nearOverlay - 1};
   width: 100%;
   transition: transform 200ms ${transitions.easing}, opacity 200ms ${transitions.easing};
-`
-
-export const ControlButton = styled.button<ControlButtonProps>`
-  margin-right: 0.5em;
-  display: flex !important;
-  padding: 0.5em;
-  cursor: pointer;
-  border: none;
-  background: none;
-  border-radius: 100%;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-  transition: background ${transitions.timings.player} ease-in !important;
-
-  & > svg {
-    filter: drop-shadow(0 1px 2px ${colors.transparentBlack[32]});
-    width: 1.5em;
-    height: 1.5em;
-  }
-
-  :hover {
-    background-color: ${colors.transparentPrimary[18]};
-    backdrop-filter: blur(${sizes(8)});
-    transition: none !important;
-
-    ${() => ControlButtonTooltip} {
-      transition: none !important;
-      opacity: ${({ showTooltipOnlyOnFocus }) => (showTooltipOnlyOnFocus ? 0 : 1)};
-    }
-  }
-
-  :active {
-    background-color: ${colors.transparentPrimary[10]};
-  }
-
-  :focus {
-    ${() => ControlButtonTooltip} {
-      /* turn off transition on mouse enter, but turn on on mouse leave */
-      transition: none !important;
-      opacity: ${({ disableFocus }) => (disableFocus ? 0 : 1)};
-    }
-  }
-
-  :focus-visible {
-    ${() => ControlButtonTooltip} {
-      opacity: ${({ disableFocus }) => (disableFocus ? 0 : 1)};
-    }
-  }
-
-  :focus:not(:focus-visible):hover {
-    ${() => ControlButtonTooltip} {
-      transition: none !important;
-      opacity: ${({ showTooltipOnlyOnFocus }) => (showTooltipOnlyOnFocus ? 0 : 1)};
-    }
-  }
-
-  :focus:not(:focus-visible):not(:hover) {
-    ${() => ControlButtonTooltip} {
-      opacity: 0;
-    }
-  }
-
-  ${focusStyles}
-`
-
-type ControlButtonTooltipProps = {
-  tooltipPosition?: 'left' | 'right'
-  showTooltipOnlyOnFocus?: boolean
-}
-
-export const ControlButtonTooltip = styled.div<ControlButtonTooltipProps>`
-  ${({ tooltipPosition }) => tooltipPosition === 'left' && 'left: 0'};
-  ${({ tooltipPosition }) => tooltipPosition === 'right' && 'right: 0'};
-
-  opacity: 0;
-  pointer-events: none;
-  position: absolute;
-  bottom: calc(3em);
-  background: ${colors.transparentBlack[54]};
-  backdrop-filter: blur(${sizes(8)});
-  display: flex;
-  align-items: center;
-  padding: 0.5em;
-  white-space: nowrap;
-  transition: opacity ${transitions.timings.player} ease-in, backdrop-filter ${transitions.timings.player} ease-in !important;
-`
-
-export const ControlButtonTooltipText = styled(Text)`
-  /* 12px */
-  font-size: 0.75em;
 `
 
 export const VolumeSliderContainer = styled.div`
@@ -216,7 +123,7 @@ export const VolumeControl = styled.div`
     }
   }
 `
-export const VolumeButton = styled(ControlButton)`
+export const VolumeButton = styled(PlayerControlButton)`
   cursor: pointer;
   margin-right: 0;
   display: flex;
